Fixed search modal and nav on mobile/desktop and fixed bug on ipad view with leaderboard on home page
