+++ src/TTX.Web/routes/(_islands)/MobileNav.tsx
@@ -1,44 +1,21 @@
-import { useSignal } from "@preact/signals";
-import Search from "./Search.tsx";
+import { Signal } from "@preact/signals";
 import { State } from "../../utils.ts";
 
 export default function MobileNav(
-  { urls, state, url }: {
-    urls: { url: string; label: string }[];
+  { urls, state, url, isSearchOpen, isMobileMenuOpen, toggleMobileMenu }: {
+    urls: readonly { url: string; label: string }[];
     state: State;
     url: URL;
+    isSearchOpen: Signal<boolean>,
+    isMobileMenuOpen: Signal<boolean>
+    toggleMobileMenu: () => void
   },
 ) {
   const from = encodeURIComponent(url.pathname);
   const user = state.user;
-  const isMobileMenuOpen = useSignal(false);
-  const toggleMobileMenu = () => {
-    isMobileMenuOpen.value = !isMobileMenuOpen.value;
-  };
 
   return (
     <>
-      <button
-        onClick={toggleMobileMenu}
-        type="button"
-        aria-label="Toggle menu"
-        class="btn btn-ghost lg:hidden"
-      >
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          class="h-5 w-5"
-          fill="none"
-          viewBox="0 0 24 24"
-          stroke="currentColor"
-        >
-          <path
-            stroke-linecap="round"
-            stroke-linejoin="round"
-            stroke-width="2"
-            d="M4 6h16M4 12h8m-8 6h16"
-          />
-        </svg>
-      </button>
       {isMobileMenuOpen.value && (
         <>
           <button
@@ -48,7 +25,7 @@ export default function MobileNav(
             aria-label="Close menu"
           >
           </button>
-          <div class="fixed left-0 top-0 z-50 h-screen w-64 bg-white shadow-lg backdrop-blur backdrop-contrast-100 backdrop-saturate-200 backdrop-filter transition-all duration-500 ease-in-out lg:hidden dark:bg-black">
+          <div class="fixed left-0 top-0 z-50 h-screen w-full bg-white shadow-lg backdrop-blur backdrop-contrast-100 backdrop-saturate-200 backdrop-filter transition-all duration-500 ease-in-out lg:hidden dark:bg-black">
             <div class="mt-4 flex h-full flex-col">
               <div class="flex items-center justify-between p-4">
                 <a
@@ -94,7 +71,29 @@ export default function MobileNav(
                     </li>
                   ))}
                   <li>
-                    <Search state={state} />
+                    <button
+                      type="button"
+                      onClick={() => {
+                        isSearchOpen.value = true;
+                      }}
+                      class="block rounded-lg px-4 py-2 text-lg font-medium underline-offset-2 hover:bg-gray-100 flex items-center justify-center gap-2"
+                    >
+                      Search
+                      <svg
+                        xmlns="http://www.w3.org/2000/svg"
+                        fill="none"
+                        viewBox="0 0 24 24"
+                        stroke-width="1.5"
+                        stroke="currentColor"
+                        class="size-5"
+                      >
+                        <path
+                          stroke-linecap="round"
+                          stroke-linejoin="round"
+                          d="m21 21-5.197-5.197m0 0A7.5 7.5 0 1 0 5.196 5.196a7.5 7.5 0 0 0 10.607 10.607Z"
+                        />
+                      </svg>
+                    </button>
                   </li>
                 </ul>
               </nav>
