+++ src/TTX.Web/routes/(_islands)/Search.tsx
@@ -1,4 +1,4 @@
-import { useSignal } from "@preact/signals";
+import { Signal, useSignal } from "@preact/signals";
 import { debounce } from "@std/async/debounce";
 import Modal from "@/islands/Modal.tsx";
 import { useRef } from "preact/hooks";
@@ -6,8 +6,7 @@ import { getRecentCreators, SearchResult } from "../../lib/search.ts";
 import { getApiClient } from "../../lib/index.ts";
 import { State } from "../../utils.ts";
 
-export default function Search({ state }: { state: State }) {
-  const isSearchOpen = useSignal(false);
+export default function SearchModal({ state, isSearchOpen }: { state: State, isSearchOpen: Signal<boolean> }) {
   const apiClient = getApiClient(state.token);
   const recent = getRecentCreators();
   const result = useSignal<SearchResult[]>([]);
@@ -108,34 +107,11 @@ export default function Search({ state }: { state: State }) {
 
   return (
     <>
-      <button
-        type="button"
-        onClick={() => {
-          isSearchOpen.value = true;
-        }}
-        class="relative z-10 rounded-2xl"
-      >
-        Search
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          fill="none"
-          viewBox="0 0 24 24"
-          stroke-width="1.5"
-          stroke="currentColor"
-          class="size-5"
-        >
-          <path
-            stroke-linecap="round"
-            stroke-linejoin="round"
-            d="m21 21-5.197-5.197m0 0A7.5 7.5 0 1 0 5.196 5.196a7.5 7.5 0 0 0 10.607 10.607Z"
-          />
-        </svg>
-      </button>
       <Modal isOpen={isSearchOpen.value}>
         <button type="button" onClick={close} aria-label="close" className="modal-backdrop size-full absolute">
-          <div class="fixed inset-0 bg-black/10 backdrop-blur-sm"></div>
+          <div class="fixed inset-0 bg-black/10 backdrop-blur-sm size-full"></div>
         </button>
-        <div class="modal-box bg-base-200 cursor-default z-10 flex h-[28rem] w-full max-w-md flex-col rounded-xl p-6 shadow-2xl shadow-purple-500/20">
+        <div class="modal-box bg-base-200/50 cursor-default z-10 flex h-[28rem] w-11/12 md:w-full max-w-md flex-col rounded-xl p-6 shadow-2xl shadow-purple-500/20">
           <div class="join flex">
             <input
               type="text"
